+++ spec/scenario/config/initializers/castronaut.rb
@@ -1,1 +1,0 @@
-Castronaut.config.connect_activerecord