+++ spec/scenario/config/initializers/devise.rb
@@ -1,4 +1,4 @@
 Devise.setup do |config|
   require "devise/orm/active_record"
-  config.timeout_in = 7200
+  config.timeout_in = 7200.seconds
 end