+++ spec/scenario/db/migrate/20100401102949_create_tables.rb
@@ -1,17 +1,17 @@
-class CreateTables < ActiveRecord::Migration
+class CreateTables < ActiveRecord::Migration[4.2]
   def self.up
     create_table :users do |t|
       t.string :username, :null => false
       t.datetime :remember_created_at
       t.string :email
-      
+
       # trackable
       t.integer :sign_in_count
       t.datetime :current_sign_in_at
       t.datetime :last_sign_in_at
       t.string :current_sign_in_ip
       t.string :last_sign_in_ip
-      
+
       t.timestamps
     end